packages/extractors/pyproject.toml
[project]
name = "ledger-lift-extractors"
version = "0.1.0"
description = "Extraction helpers (placeholder)"
requires-python = ">=3.11"
dependencies = []

[project.optional-dependencies]
consensus = [
    "pdfplumber==0.10.3",
    "camelot-py[base]==0.10.1",
    "tabula-py==2.8.2",
    "pandas==2.1.4",
    "numpy==1.24.4",
]
